Changing lastPKIStatus from int to long in cmp_int.h
as it is always set through functions returning long

Same for refNum locally used in CMP_CERTREPMESSAGE_get_certificate()

+++ crypto/cmp/cmp_int.h
@@ -97,7 +97,7 @@ struct cmp_ctx_st {
     int maxPollTime;
     /* PKIStatus of last received IP/CP/KUP */
     /* TODO: this should be a stack since there could be more than one */
-    int lastPKIStatus;
+    long lastPKIStatus;
     /* failInfoCode of last received IP/CP/KUP */
     /* TODO: this should be a stack since there could be more than one */
     unsigned long failInfoCode;
